Tin処理を別モジュールに分離
